+++ cmd/tt/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/DragoHex/task-tracker/pkg/lib"
+
+func main() {
+	lib.Root()
+}
