+++ pkg/lib/task.go
@@ -1,4 +1,4 @@
-package main
+package lib
 
 import (
 	"encoding/json"
